requirements.txt
torch
torchvision
matplotlib
scikit-learn
salesforce-lavis
munkres
clingo
lightning-flash